backend/requirements.txt
# LiveHub FastAPI Backend Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# JWT Authentication
python-jose[cryptography]>=3.3.0

# File uploads
python-multipart>=0.0.9

# Object Storage
minio>=7.2.0

# Vector Database
qdrant-client>=1.7.0

# AI - Face Detection (DeepFace + TensorFlow)
deepface>=0.0.79
tensorflow>=2.15.0
tf-keras>=2.15.0
opencv-python>=4.9.0
numpy>=1.24.0,<2.0.0

# Background Tasks (using built-in ThreadPoolExecutor, no external dependencies needed)
redis>=5.0.0

# Image Processing
pillow>=10.0.0

# HTTP Client
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
